+++ src/Widgets/Concerns/CalculatesContextualColors.php
@@ -63,7 +63,7 @@ trait CalculatesContextualColors
         }
 
         if (is_callable($definition)) {
-            $rawValue = $definition($value);
+            $rawValue = $definition($model, $value);
             return is_string($rawValue) ? $this->fromColorString($rawValue) : new ColorAttributes($rawValue, null);
         }