+++ Changelog.md
@@ -1,5 +1,11 @@
 # AppShell Changelog
 
+## Unreleased
+##### 2025-XX-YY
+
+- Added closure-based interpolation support to Text widget attributes like class, style and title
+- Changed the closure parameter passed to the contextual color calculation to be the "model, pre-rendered text" - earlier it was the pre-rendered text only
+
 ## 4.15.0
 ##### 2025-12-08